requirements.txt
amqp==5.1.1
asgiref==3.5.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
Django==4.0
django-smart-selects==1.5.9
djangorestframework==3.13.1
elasticsearch==7.17.4
elasticsearch-dsl==7.4.0
kombu==5.2.4
Pillow==9.2.0
prompt-toolkit==3.0.31
psycopg2==2.9.3
python-dateutil==2.8.2
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
